Added pickle interpreter
This initial commit of the pickle interpreter adds a component and
example code that can invoke the pickle interpreter. It needs greater
integration in the project, which would include (but not necessarily be
limited to); functionality for running the interpreter in its own
thread, functionality for writing/reading/executing files from
non-volatile storage, controlling the WiFi module, RTOS stuff and more.

Other things such as project layout will also need to be fixed/improved
upon.